# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceKernel API",
    "version": "1.0.0",
    "description": "The voice infrastructure layer for regulated enterprise. Create agents, place calls, ground answers in your knowledge, and stream every event into your own systems."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/provider/{path}": {
      "get": {
        "tags": [
          "Provider passthrough"
        ],
        "summary": "Mediated access to any of the 97 provider operations",
        "description": "Same auth, tenant isolation, rate limiting and audit as native routes. GET /v1/provider/_operations lists everything reachable.",
        "responses": {
          "200": {
            "description": "The upstream response."
          }
        }
      },
      "parameters": [
        {
          "name": "path",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Provider path, e.g. \"assistant\" or \"call/abc123\"."
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "vk_live_…"
      }
    },
...
  }
}